fix: introduce CI failure check
Our CI matrix is dynamic so it cannot be used for required status checks in github. However we can add a failure step that can get triggered if any dependency fails or is cancelled. If this is the case, this job will fail. If CI is successful, the CI Completion Check will be skipped.