fix: make sure that CTFs do not write a global access into the local blob
Currently local resources added into a CV return a resource that has the globalAccess field set always. However, because CTFs are local archives that global reference contains the virtual `ctf.ocm.software` address as a reference which is not global. That means we should ONLY write the global access if we are working in a remote repository, otherwise we should only write the localBlob with its local reference (which can always be resolved if you know where the component version is).