fix: fixup input contract
this changes the input method contract:

```
// ResourceInputMethod is the interface for processing a resource with an input method decaled as per
// [spec.Resource.Input].
//
// The method will get called with the raw specification for the resource and is expected
// to return a runtime resource specification with [descriptor.Resource.Access] set to the
// now resulting access type.
//
// For Input Methods wishing to colocate their uploaded resources with the component version
// they are part of, the targeted [ComponentVersionRepository] is passed in via the Options struct,
// and can be used to upload the resource into the OCM component version repository.
type ResourceInputMethod interface {
	ProcessResource(ctx context.Context, resource *spec.Resource, opts Options) (processed *descriptor.Resource, err error)
}
```

This makes it so that every method can decide how to deal with the specified resource. This resource will be guaranteed to have an input method defined and that means that each method can decide to use a local component version target repository, or to initialize its own type of repository, e.g. as a resource repository.