fix: ensure that resources by reference have their relation and version defaulted
previously resources by reference did not have their version and relation defaulted when they were missing.

This makes it so that resources referenced externally (via accessType instead of input type) get a default of external relation, as they are assumed to be not maintained by the component creator.

At the same time, if the resource is not versioned, we assume that the resource is versioned alongside the component and therefore the component version can be used as defaulting value.