feat!: allow specifying a copy option in the constructor in favor of `--copy-resources`
this makes it so that resources and sources have a new attribute called `copyPolicy` that allows to specify if the resource should be copied by value or not. This attribute only is valid right now if the specification of the given element is an access, NOT an input.

Now we can separate per resource, if the resource should be added by value based on the specification.

Note that this now means that we no longer have an option to specify via callback if the resource is copied by value. this can be decided while building the constructor specification.