fix: remove flaky error write and ensure benchmarks are run
currently there were some benchmark tests not run regularly. this is not good.

i have changed this so the benchmarks are run and then noticed that one of them flaked because in repository resolution we do not use the correct error variable every time.

also there was a log message written every time a repo resolution failure occurs. however we dont care about that if we have at least one success.

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>