fix: optimize layout copying
This changes the underlying CopyToOCILayoutInMemory implementation so that any call to it triggers a layout copy asynchronously. This omits the buffer and only causes buffering at the usage state if the digest or size are requested.

This effectively makes it so that the blobs returned only compute size and digest when needed, not by default

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>