feat: direct blob
this implements a direct blob, aka a blob that can be read from many times independently but that does not buffer itself.

this is especially useful when forwarding already existing buffers such as bytes.Buffer or plain []byte slices.

One caveat towards the existing "inmemory" implementation is that because it does not buffer itself, we cannot know the digest and if someone wants to compute it, they will have to buffer the data themselves and compute it dynamically.

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>