feat: support stateful, serialized io.Reader's in direct blobs
when we don't have io.ReaderAt backed implementations of blobs (for example when working with io.Pipe which is not resettable), then we need to base our implementation solely on io.Reader. However the blobs ReadCloser() methods can in theory be concurrently called, so we need a serializing blob that is able to protect from bad reads.

This fixes this problem and allows direct blobs to receive an io.Reader natively, switching to serialized access under the hood if needed.

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>