chore(cli): restructure integration module path
### What this commit does:
- Renamed `cli/cmd/download/plugin/integration` to `cli/integration`.
- Updated `go.mod` to reflect the new module path.
- Adjusted `Taskfile.yml` entries to align with the new directory structure.
- Incremented `configuration` dependency version from `v0.0.6` to `v0.0.7`.

We need this because otherwise the integration test is not picked up by the CI (WIP on how we fix this permanently)

### Why this is needed:
This restructuring simplifies the module structure by placing integration tests in a more logical and concise directory, improving maintainability. Additionally, the dependency update ensures the latest improvements and fixes are applied.

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>