fix(fetch): streamline blob retrieval by always referencing the digest instead of the annotations
we can significantly streamline reading of local blobs by using the local reference from the access instead of its identity annotation.

the added benefit of this is that we will now be able to read OCM OCI descriptors even if they do not have these annotations set. This is possible because we now can directly use the reference digest. This makes us compatible with any OCM component version created without these (not standard, but created as such by old versions of CLI and other tooling)

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>