fix(linter): do not verify the config on each matrix run but do it once in the begin (#1220)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

Our linter is unstable because it verifies the config on EACH RUN which
is plenty and effectively times out at some point because of it.

Instead of verifying on each run, we disable verify in the linter action
and run it once before the linter starts instead.


https://github.com/open-component-model/open-component-model/actions/runs/19440119253/job/55621133428

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

Signed-off-by: Gergely Brautigam <182850+Skarlso@users.noreply.github.com>