feat(scheme): expose scheme retrieval methods for internal plugins (#1268)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
Added methods to retrieve the runtime scheme in multiple components,
preparing for simpler internal plugin registration.

If you feel like you are lacking context for some of the changes in this
PR, it is part of this larger
[PR](https://github.com/open-component-model/open-component-model/pull/1228).

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
Contributes to
https://github.com/open-component-model/ocm-project/issues/761

---------

Signed-off-by: Fabian Burth <fabian.burth@sap.com>