feat: add root component for plugin registry (#1229)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

Fixes https://github.com/open-component-model/ocm-project/issues/598.

This is the current flow:

- either a manual run triggers the workflow or a plugin update workflow
once it finishes remote triggers this workflow with its newly published
version and name.
- once run, it will check if the root registry component exists or not.
- if it doesn't exist it will create it current using the template under
.github/config/plugin-registry-constructor.yaml
- ( why template? because it's easy and visible and version controlled
what kind of plugins there are, I didn't want to construct it
dynamically, because I'm afraid that we will have no idea what plugins
there are ( other then doing a component get and checking. but I'm up
for a debate on this )
- based on if a new plugin is added or an existing plugin is being
modified it will automatically calculate the new version and bump it
accordingly. new plugin -> minor update, existing plugin -> patch
update. Major updates should be performed by a person pushing a new
style of the plugin component or some such
- the new version is constructed and pushed

Right now, there is no hashing of the component reference or the plugin
registry. This will be added in a follow-up.

Flow:
Release new Go Submodule for Helm -> Triggers a publish on the plugin ->
Triggers an update of the root component update.

TODO/Questions: 

- Right now, the plugin checking itself and the updates and such are
agnostic of the plugin's name and such that is being pushed. However,
this presents and interesting problem. Do we just blindly add plugins
all the time and the template is literally just the main part with an
empty componentReference? Or do we _restrict_ what plugins we know about
by templating that part like it is right now, we DON'T allow pushing a
plugin that does NOT exist like the above flow thus currently

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

---------

Signed-off-by: Gergely Brautigam <182850+Skarlso@users.noreply.github.com>