feat(registry): simplify internal plugin registration methods (#1271)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
Refactor the registration methods for internal plugins to use a more
consistent and streamlined approach. This change enhances readability
and maintainability by reducing redundancy in the code.

This is what the changes in this previous
[PR](https://github.com/open-component-model/open-component-model/pull/1268)
were needed for.

**Notes for the Reviewers**
The changes are quite repetitive. But pay special attention to the
registration method of input.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
Contributes to
https://github.com/open-component-model/ocm-project/issues/761

---------

Signed-off-by: Fabian Burth <fabian.burth@sap.com>
Co-authored-by: Jakob Möller <jakob.moeller@sap.com>