feat: add an option to specify schemes for the repository provider (#1282)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

Currently, a new Component Version Provider is created with a hard-coded
scheme, [see
here](https://github.com/open-component-model/open-component-model/blob/main/bindings/go/oci/repository/provider/provider.go#L91).
The scheme is a global variable created
[here](https://github.com/open-component-model/open-component-model/blob/main/bindings/go/oci/spec/repository/scheme.go#L12).

However, the global variable does not register the legacy schemes which
is why we encountered a problem on the component controller.

This change provides the opportunity to define a custom scheme and pass
it to the repository provider.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

---------

Signed-off-by: Frederic Wilhelm <frederic.wilhelm@sap.com>