chore(plugin): make plugin module compatible with latest version of runtime module (#1320)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
The latest version of the runtime module deleted the
GenerateJSONSchemaForType method. We will move away from it eventually.
But to be able to make these changes incrementally, we re-implement it
in the plugin module in the meantime.

**Notes for the Reviewers**
The JSON Schema code is a 1:1 copy from the previous runtime module
version.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
Contributes to
https://github.com/open-component-model/ocm-project/issues/761

---------

Signed-off-by: Fabian Burth <fabian.burth@sap.com>
Co-authored-by: Jakob Möller <jakob.moeller@sap.com>