fix(oci): make consumer identity for ctf return an error (#1323)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
After the changes in #1191, the credential resolution is broken as we
attempt to resolve credentials with the consumer identity returned for
ctf from dockerconfig credential repository in the fallback logic to any
consumer identity type of our credential resolution.

This is a preliminary fix for the issue, unblocking us. A broader scoped
issue will be created as follow up.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
Contributes to
https://github.com/open-component-model/ocm-project/issues/737

---------

Signed-off-by: Fabian Burth <fabian.burth@sap.com>