chore(deps): update helm plugin to refactored plugin manager (#1315)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
During the plugin manager refactoring, constants were moved to different
packages. This updates the helm plugin to import the updated packages.

**Notes for the Reviewer**
I couldn't update the `runtime` module to 0.0.4 yet, as the plugin
module will have to be updated to the JSON schema generator introduced
[here](https://github.com/open-component-model/open-component-model/pull/1219).

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
Contributes to
https://github.com/open-component-model/ocm-project/issues/760

---------

Signed-off-by: Fabian Burth <fabian.burth@sap.com>
Co-authored-by: Matthias Bruns <github@matthiasbruns.com>