fix: make sure prototypes can be used for reregistration (#1284)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

Previously, in the scheme we did not store by Prototype but by type.
Because the scheme did not key by Prototype, one could register multiple
defaults with the same underlying prototype value. However, this is
incorrect. Instead there should be only one prototype instance at any
given time in a scheme. If there are multiple registrations happening in
succession, the prototype should still stay atomic, and all calls should
instead only append to the aliases

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

was discovered as part of scheme debugging in
https://github.com/open-component-model/open-component-model/pull/1188/commits/e25a7d0cd6041ccf9f707d86fc1ac59158fd516f

---------

Signed-off-by: Jakob Möller <contact@jakob-moeller.com>