feat(cel): schema based resource parser
when we actually do have a schema, it is useful to use the schema for static analysis purposes. This is a first opinionated implementation of this parsing logic that also works based on the JSON Schema and not just on the resource itself.

NOTE: this has many opinionations and limits builtin and is purpose-built for our runtime. We will extend this on demand later.
Signed-off-by: Jakob Möller <contact@jakob-moeller.com>