fix: rm nested fields and descriptor scheme (#1391)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

this makes it so that we can override the scheme that is used and
referenced for a type with an annotation. we use this for descriptor v2
to embed the descriptor scheme which (for now) is pregenerated and
maintained manually instead of autogenerated.

It also removes nested ids to embed the schemes properly so that they
are considered part of their parent scheme and def resolution works as
we intended it.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

---------

Signed-off-by: Jakob Möller <contact@jakob-moeller.com>