feat(cel): schema based resource parser (#1378)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

when we actually do have a schema, we should use the schema for static
analysis purposes. This is a first opinionated implementation of this
parsing logic that also works based on the JSON Schema and not just on
the resource itself.

It also contributes some fixes of code that I pushed in as part of our
cel implementation that was not yet complete. As before, this is
inspired by KRO but differs in that we use proper JSON Schema for our
backing types.

NOTE: this has many opinionations and limits builtin and is
purpose-built for our runtime. We will extend this on demand later.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

This is the actual static parsing that happens based on the field and is
used together with our expression engine to get a static evaluation
result for any transformation. Part of
https://github.com/open-component-model/ocm-project/issues/536

---------

Signed-off-by: Jakob Möller <contact@jakob-moeller.com>
Co-authored-by: Gergely Brautigam <182850+skarlso@users.noreply.github.com>