chore: remove issue templates to centralize issue template management to `.github` repo (#1434)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
We decided we want to extend our issue templates to provide stronger
guidelines and guardrails about required information. We decided to
remove the issue templates from this repository completely and manage
the issue templates centrally in the `.github` repository as described
in
https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/creating-a-default-community-health-file.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

Signed-off-by: Fabian Burth <fabian.burth@sap.com>