fix: add missing `$id` properties to JSON schema definitions (#1449)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

For external schemas, because we now deduplicate them, we can go ahead
and preserve the ID. That will ensure nested definitions are preserved

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

This relates to our recent refactoring of the schema generator. If we
wouldnt do this, jsonschema resolution for the embedded component
descriptor would fail

Signed-off-by: Jakob Möller <contact@jakob-moeller.com>