chore: error on attempted runtime descriptor serialization (#1448)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
Runtime Descriptors should never be used for serialization. Attempting
to do so silently did not work as expected so far, due to the json
annotations. This could easily lead to confusion.

This PR introduces errors on attempting to marshal or unmarshal a
runtime descriptor.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

---------

Signed-off-by: Fabian Burth <fabian.burth@sap.com>
Co-authored-by: Frederic Wilhelm <fre.wilhelm@gmail.com>
Co-authored-by: Frederic Wilhelm <frederic.wilhelm@sap.com>