chore: remove replication controller (#1482)
As the current design of the replication controller is based on the ocm
lib v1 and will have breaking changes when migrating to ocm lib v2, it
is removed for now. By this, we will get rid of the ocm lib v1
dependency hell quicker and do not have to maintain the
replication-tests that will change either way.

All the related files are removed because that is simpler than disabling
it everywhere.

It will be re-implemented with
https://github.com/open-component-model/ocm-project/issues/631

Signed-off-by: Frederic Wilhelm <frederic.wilhelm@sap.com>