fix: improve error handling and logging in credential resolution for resource digest processing (#1514)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

makes it so that credentials are resolved and checked with our common
helper in construct, and makes it so that that credential helper has the
ability to avoid failing if there are no credentials present.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

needed to allow correct digest processing of resources with an access
that can but not must have credentials.

relates to
https://github.com/open-component-model/ocm-project/issues/790

Signed-off-by: Jakob Möller <contact@jakob-moeller.com>